se crea la interfaz de isuario para mostrar el album y sus canciones

+++ app/src/main/java/com/bancosantander/globallogicbryanescalona/data/remote/model/SongListEntry.kt
@@ -1,12 +1,13 @@
 package com.bancosantander.globallogicbryanescalona.data.remote.model
 
+import android.util.Log
 import com.bancosantander.globallogicbryanescalona.domain.model.SongList
 
 data class SongListEntry (
     val wrapperType:String,
     val kind:String,
-    val artistId:Int,
-    val collectionId:Int,
+    val artistId:Long,
+    val collectionId:Long,
     val trackId:Int,
     val artistName:String,
     val collectionName:String,
